requirements.txt
pyyaml>=5.1
pytest>=6.0.0
pytest-mock>=3.6.0